+++ src/twin/mod_test.rs
@@ -2,6 +2,7 @@
 mod mod_test {
     use super::super::*;
     use crate::test_util::Testrunner;
+    use serde_json::json;
     use std::env;
     use std::sync::mpsc;
     use std::sync::mpsc::TryRecvError;
@@ -236,4 +237,72 @@ mod mod_test {
 
         // ToDo: add more tests based on network adapters available on build server?
     }
+
+    #[test]
+    fn factory_reset_test() {
+        assert!(factory_reset::reset_to_factory_settings(json!({
+            "type": 1,
+            "restore_settings": ["wifi"]
+        }),)
+        .unwrap_err()
+        .to_string()
+        .starts_with("No such file or directory"));
+
+        assert!(factory_reset::reset_to_factory_settings(json!({
+            "type": 1,
+        }),)
+        .unwrap_err()
+        .to_string()
+        .starts_with("No such file or directory"));
+
+        assert_eq!(
+            factory_reset::reset_to_factory_settings(json!({
+                "restore_settings": ["wifi"]
+            }),)
+            .unwrap_err()
+            .to_string(),
+            "reset type missing or not supported"
+        );
+
+        assert_eq!(
+            factory_reset::reset_to_factory_settings(json!({
+                "type": 1,
+                "restore_settings": ["unknown"]
+            }),)
+            .unwrap_err()
+            .to_string(),
+            "unknown restore setting received"
+        );
+    }
+
+    #[test]
+    fn user_consent_test() {
+        assert!(consent::user_consent(json!({"swupdate": "1.0.0"}))
+            .unwrap_err()
+            .to_string()
+            .starts_with("user_consent: open user_consent.json for write"));
+
+        assert_eq!(
+            consent::user_consent(json!({})).unwrap_err().to_string(),
+            "unexpected parameter format"
+        );
+
+        assert_eq!(
+            consent::user_consent(json!({"swupdate": "1.0.0", "another_component": "1.2.3"}))
+                .unwrap_err()
+                .to_string(),
+            "unexpected parameter format"
+        );
+
+        let tr = Testrunner::new(function_name!().split("::").last().unwrap());
+        tr.copy_file("testfiles/consent_conf.json");
+
+        env::set_var("CONSENT_DIR_PATH", tr.get_dirpath().as_str());
+
+        tr.copy_directory("testfiles/test_component");
+
+        assert!(consent::user_consent(json!({"test_component": "1.0.0"})).is_ok());
+    }
+
+    //ToDo: add remaining unittests
 }