+++ src/twin/network_status.rs
@@ -1,0 +1,127 @@
+use super::Twin;
+use crate::{twin, ReportProperty};
+use anyhow::{Context, Result};
+use log::{error, info};
+use network_interface::{Addr, NetworkInterface, NetworkInterfaceConfig};
+use serde::Serialize;
+use serde_json::json;
+use serde_with::skip_serializing_none;
+use std::collections::HashMap;
+
+pub fn refresh_network_status(_in_json: serde_json::Value) -> Result<Option<serde_json::Value>> {
+    info!("network status requested");
+
+    twin::get_or_init(None).report(&ReportProperty::NetworkStatus)?;
+
+    Ok(None)
+}
+
+impl Twin {
+    pub fn update_include_network_filter(
+        &mut self,
+        include_network_filter: Option<&Vec<serde_json::Value>>,
+    ) -> Result<()> {
+        if include_network_filter.is_none() {
+            self.include_network_filter.take();
+            return self
+                .report_impl(json!({ "network_interfaces": json!(null) }))
+                .context("report_network_status: report_impl");
+        }
+
+        let mut new_include_network_filter: Vec<String> = include_network_filter
+            .unwrap()
+            .iter()
+            .filter(|e| {
+                if !e.is_string() {
+                    error!(
+                        "unexpected format in desired include_network_filter. ignore: {}",
+                        e.to_string()
+                    );
+                }
+                e.is_string()
+            })
+            .map(|e| e.as_str().unwrap().to_string().to_lowercase())
+            .collect();
+
+        // enforce entries only exist once
+        new_include_network_filter.sort();
+        new_include_network_filter.dedup();
+
+        // check if desired include_network_filter changed
+        if self.include_network_filter.is_none()
+            || self
+                .include_network_filter
+                .as_ref()
+                .unwrap()
+                .ne(&new_include_network_filter)
+        {
+            self.include_network_filter
+                .replace(new_include_network_filter);
+            self.report_network_status()
+        } else {
+            info!("desired include_network_filter didn't change");
+            Ok(())
+        }
+    }
+
+    pub fn report_network_status(&mut self) -> Result<()> {
+        #[skip_serializing_none]
+        #[derive(Serialize)]
+        struct NetworkReport {
+            #[serde(default)]
+            name: String,
+            mac: String,
+            addr_v4: Option<Vec<String>>,
+            addr_v6: Option<Vec<String>>,
+        }
+
+        let mut interfaces: HashMap<String, NetworkReport> = HashMap::new();
+
+        NetworkInterface::show()
+            .context("report_network_status")?
+            .iter()
+            .filter(|i| {
+                self.include_network_filter
+                    .as_ref()
+                    .unwrap()
+                    .iter()
+                    .any(|f| {
+                        let name = i.name.to_lowercase();
+                        match (f.starts_with("*"), f.ends_with("*"), f.len()) {
+                            (_, _, 0) => false,                                     // ""
+                            (a, b, 1) if a || b => true,                            // "*"
+                            (true, true, len) => name.contains(&f[1..len - 1]),     // ""*...*"
+                            (true, false, len) => name.ends_with(&f[1..len]),       // "*..."
+                            (false, true, len) => name.starts_with(&f[0..len - 1]), // "...*"
+                            _ => name.eq(f),                                        // "..."
+                        }
+                    })
+            })
+            .for_each(|i| {
+                let entry = interfaces.entry(i.name.clone()).or_insert(NetworkReport {
+                    addr_v4: None,
+                    addr_v6: None,
+                    mac: i.mac_addr.clone().unwrap_or_else(|| "none".to_string()),
+                    name: i.name.clone(),
+                });
+
+                match i.addr {
+                    Some(Addr::V4(addr)) => entry
+                        .addr_v4
+                        .get_or_insert(vec![])
+                        .push(addr.ip.to_string()),
+                    Some(Addr::V6(addr)) => entry
+                        .addr_v6
+                        .get_or_insert(vec![])
+                        .push(addr.ip.to_string()),
+                    None => error!("report_network_status: ip address is missing"),
+                };
+            });
+
+        self.report_impl(json!({
+            "network_interfaces": json!(interfaces.into_values().collect::<Vec<NetworkReport>>())
+        }))
+        .context("report_network_status")
+        .map_err(|err| err.into())
+    }
+}